% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualBounds.R
\name{actualBounds}
\alias{actualBounds}
\title{Actual Rejection Boundary Based on Actual Number of Events in Group Sequential Design}
\usage{
actualBounds(
  planned.events = c(126, 210),
  act.events = c(140, 250),
  sf = gsDesign::sfLDOF,
  alpha = 0.025
)
}
\arguments{
\item{planned.events}{A vector of planned target events for IAs and FA}

\item{sf}{Alpha spending function. Default sf = gsDesign::sfLDOF, i.e.,
LanDeMets implementation of O'Brien Fleming spending function.}

\item{alpha}{Overall alpha for the test in the group sequential design.
Default, alpha = 0.025. Alpha must be one-sided.}

\item{actual.events}{A vector of actual target events for IAs and FA}
}
\value{
An object with values
\itemize{
\item planned.p Planned rejection bounds in p value (1-sided) based on planned number of events at each analysis
\item planned.z Planned rejection bounds in z value  based on planned number of events at each analysis
\item actual.p Actual rejection bounds in p value (1-sided) based on actual number of events at each analysis and the FA bound is adjusted by the alpha already spent at IAs and the actual number of events at FA
\item actual.z Actual rejection bounds in z value  based on actual number of events at each analysis  and the FA bound is adjusted by the alpha already spent at IAs and the actual number of events at FA
}
}
\description{
This function calculates the actual rejection boundary at FA based on the
alpha spending already occurred at interim analyses. The alpha spending for
interim analyses are according to the information fractions relative to
the planned target events at final analysis. The rejection boundary at
the final analysis is calculated according to the alpha already spent at IAs
and the actual number of events observed at final analysis.
}
\examples{

actualBounds(planned.events=c(126, 210), act.events=c(140, 250), sf=gsDesign::sfLDOF, alpha=0.025)

}
