% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrEvents.R
\name{corrEvents}
\alias{corrEvents}
\title{Expected Number of Events Over Time With Staggered Entry For Overlapping Populations}
\usage{
corrEvents(
  T = 24,
  n = list(AandB = 300, AnotB = 0, BnotA = 450),
  r = list(AandB = 1/2, AnotB = 0, BnotA = 1/2),
  h0 = list(AandB = function(t) {     log(2)/12 }, AnotB = function(t) {     log(2)/12
    }, BnotA = function(t) {     log(2)/12 }),
  S0 = list(AandB = function(t) {     exp(-log(2)/12 * t) }, AnotB = function(t) {    
    exp(-log(2)/12 * t) }, BnotA = function(t) {     exp(-log(2)/12 * t) }),
  h1 = list(AandB = function(t) {     log(2)/12 * 0.7 }, AnotB = function(t) {    
    log(2)/12 * 0.7 }, BnotA = function(t) {     log(2)/12 * 0.7 }),
  S1 = list(AandB = function(t) {     exp(-log(2)/12 * 0.7 * t) }, AnotB = function(t)
    {     exp(-log(2)/12 * 0.7 * t) }, BnotA = function(t) {     exp(-log(2)/12 * 0.7 *
    t) }),
  F.entry = function(t) {     (t/18)^1.5 * as.numeric(t <= 18) + as.numeric(t > 18) },
  G.ltfu = function(t) {     0 }
)
}
\arguments{
\item{T}{Analysis time calculated from first subject randomization date.}

\item{n}{Total sample size for two arms for subjects in both population A and B,
A not B, B not A. Default is NULL.}

\item{r}{Proportion of experimental subjects in both population A and B,
A not B, B not A. Default, 1/2 for 1:1 randomization stratified by A and B}

\item{h0}{Hazard function of control arm for subjects in both population A and B,
A not B, B not A. h0(t) = log(2)/m0 means T~exponential distribution with
median m0. For study design without considering heterogeneous effect in
strata for control arm, then specify the same h0(t) function across strata.}

\item{S0}{Survival function of control arm for subjects in both population A and B,
A not B, B not A. h0(t) = log(2)/m0 means T~exponential distribution with
median m0. For study design without considering heterogeneous effect in
strata for control arm, then specify the same S0(t) function across strata.
The density function f0(t) = h0(t) * S0(t).}

\item{h1}{Hazard function of experimental arm for subjects in both population A and B,
A not B, B not A. For study design without considering heterogeneous effect in
strata for the experimental arm, then specify the same h1(t) function across strata.}

\item{S1}{Survival function of experimental arm for subjects in both population A and B,
A not B, B not A. For study design without considering heterogeneous effect in
strata for the experimental arm, then specify the same h1(t) function across strata.}

\item{F.entry}{Distribution function of enrollment. For uniform enrollment,
F.entry(t) = (t/A) where A is the enrollment period, i.e., F.entry(t) = t/A for 0<=t<=A, and
F.entry(t) = 1 when t > A. For more general non-uniform enrollment with weight psi,
F.entry(t) = (t/A)^psi*I(0<=t<=A) + I(t>A). Default F.entry is uniform distribution function.}

\item{G.ltfu}{Distribution function of lost-to-follow-up censoring process. The observed
survival time is min(survival time, lost-to-follow-up time). For 3\\% drop-off
every year as a constant rate, then G.ltfu = 1-exp(-0.03/12*t), i.e., drop-off~exp distribution.
Default G.ltfu = 0 (no lost-to-followup)}
}
\value{
An object with dataframes below.
\describe{
\item{p.event}{Probability of event for each arm (p.event0: control group; p.event1: experimental group)}
\item{n.events}{Expected number of events}
\itemize{
\item subgroup: Label of subgroups
\item n.events0: number of events for control group
\item n.events1: number of events for experimental group
\item n.events.total: total number of events for two groups
\item n0: number of subjects in control arm
\item n1: number of subjects in experimental arm
\item n.total: total number of subjects
\item maturity0: maturity in control arm, percent of events
\item maturity1: maturity in experimental arm, percent of events
\item maturity: maturity in both arms, percent of events
}
\item{param}{Parameters specified: T, r, and n}
\item{param.fun}{f0, f1, F.entry, G.ltfu}
}
}
\description{
This function calculates the expected number of events under alternative hypothesis
at an analysis time, which is calculated from first subject in. The function
returns the expected number of events for each arm, based on the provided
enrollment distribution function and drop-off distribution if applicable.
If the total sample size is not provided, then only the corresponding probability of event
for each arm is provided. The expected number of events for each set of subjects is
calculated including in both A and B, in A not B, in B not A, and in A or B.
}
\examples{
#PD-L1+ subgroup and overall population tests. Control arm has exponential
#distribution with median 12 months in all strata. Experimental arm has
#proportional hazards with a hazard ratio of 0.7 in all strata. The randomization
#is stratified by PD-L1+ status. 1:1 randomization. 300 subjects in PD-L1+ and 
#450 total subjects in overall population. 3\\% drop-off every year. Enrollment
#period is 18 months and weight 1.5. The expected number of events at 24 months is
#calculated as
corrEvents(T = 24, n = list(AandB = 300, AnotB=0, BnotA=450), 
   r = list(AandB=1/2, AnotB =0, BnotA = 1/2), 
   h0=list(AandB=function(t){log(2)/12}, AnotB=function(t){log(2)/12}, 
           BnotA=function(t){log(2)/12}), 
   S0=list(AandB=function(t){exp(-log(2)/12*t)}, AnotB=function(t){exp(-log(2)/12*t)},
           BnotA=function(t){exp(-log(2)/12*t)}),
   h1=list(AandB=function(t){log(2)/12*0.70},    AnotB=function(t){log(2)/12*0.70},
           BnotA=function(t){log(2)/12*0.70}), 
   S1=list(AandB=function(t){exp(-log(2)/12 * 0.7 * t)},AnotB=function(t){exp(-log(2)/12 * 0.7 * t)},
           BnotA=function(t){exp(-log(2)/12 * 0.7 * t)}),
   F.entry = function(t){(t/18)^1.5*as.numeric(t <= 18) + as.numeric(t > 18)}, 
   G.ltfu = function(t){1-exp(-0.03/12*t)})


}
