% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fentry.R
\name{Fentry}
\alias{Fentry}
\title{Creation of Piecewise Enrolment Distribution}
\usage{
Fentry(t, time = 1:5, cum.p = c(0.05, 0.15, 0.35, 0.65, 1), xi = rep(1, 5))
}
\arguments{
\item{t}{Enrolment time}

\item{time}{A vector of time points to define the piecewise enrolment intervals
for example, time = c(1,2,3) defines enrolment intervals 0-1, 1-2, 2-3.}

\item{cum.p}{A vector of cumulative enrolment at each time point. It must have
the same length of time. For example, cum.p = c(0.2, 0.5, 1.0).}

\item{xi}{A vector of weights for each piecewise enrolment distribution function.
It must have the same length as time and cum.p. For example, xi = c(1,1,1)
means uniform enrollment in all intervals.}
}
\description{
This function creates a piecewise enrolment distribution to faciliate the
calculation of expected number of events.
}
\examples{

t = seq(0, 8, by=0.05)
fe1 = Fentry(t=t, time=1:5, cum.p = c(0.05, 0.15, 0.35, 0.65, 1.0),xi = rep(1,5))
fe2 = Fentry(t=t, time=1:5, cum.p = c(0.05, 0.15, 0.35, 0.65, 1.0),xi = rep(2.5,5))
plot(t, fe1, type="n", xlab="Enrollment time (month)", ylab="Enrolment Distribution Function")
lines(t, fe1, col=1, lty=1, lwd=2)  
lines(t, fe2, col=2, lty=2, lwd=2)  
legend(0, 1, c("Piecewise xi=1", "Piecewise xi=2.5"), lwd=rep(2,2), col=1:2, lty=1:2, cex=0.8, bty="n")

}
