% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallAHR.R
\name{overallAHR}
\alias{overallAHR}
\title{Hazard ratio of overall population for heterogeneous population}
\usage{
overallAHR(
  S0 = list(function(t) {     exp(-log(2)/12 * t) }, function(t) {     exp(-log(2)/12 *
    t) }),
  h0 = list(function(t) {     log(2)/12 }, function(t) {     log(2)/12 }),
  HR = c(0.65, 0.75),
  p = c(0.45, 0.55)
)
}
\arguments{
\item{S0}{Control arm survival function. Optional. The average HR is generally
robust with respect to S0. If S0 is not provided, an exponential distribution
with median 12 months will be used by default.}

\item{h0}{Control arm hazard rate functio. Optional. For exponential distribution,
h0 is a constant. If h0 is not provided, constant hazard of log(2)/12 is used.
If S0 and h0 are provided, they must be consistent to each other, ie, S0*h0=f0.}

\item{HR}{A vector of hazard ratios for subgroups.}

\item{p}{A vector of proportions for subgroups. sum(p) must be 1. HR and p
must be in the same order of subgroups.}
}
\value{
Average hazard ratio in the overall population combining all mutually
exclusive subgroups.
}
\description{
This function calculates the hazard ratio function of overall population when
there is heterogeneous effect in subgroups. In general, given the hazard ratio
for each mutually exclusive subgroup, the overall hazard ratio in the overall
population combining all subgroups is a function of survival time. This function
calculates the average hazard ratio, which is defined as the expected value of
the overall hazard ratio function, i.e, average Hazard Ratio = E(HR(t)), which
can be evaluated based on S0(t), however in common situations, it is generally
robust to the specification of S0(t).
}
\examples{
#Example 1. PD-L1+/- subgroups have HRs 0.65 and 0.75 with prevalence of 35\% and 65\% 
#respectively. The average hazard ratio of the overall population (PD-L1+/-) is estimated as 
HR.m12 = overallAHR(S0 = function(t){exp(-log(2)/12*t)}, h0=function(t){log(2)/12}, 
HR = c(0.65, 0.75), p = c(0.35, 0.65))

HR.m20 = overallAHR(S0 = function(t){exp(-log(2)/20*t)}, h0=function(t){log(2)/20}, 
HR = c(0.65, 1), p = c(0.70, 0.30))

HR.m20 = overallAHR(S0 = function(t){exp(-log(2)/20*t)}, h0=function(t){log(2)/20}, 
HR = c(0.55, 0.65, 1), p = c(0.45, 0.25, 0.30))

overallAHR(S0 = function(t){exp(-log(2)/20*t)}, h0=function(t){log(2)/20}, 
HR = c(0.65, 0.8), p = c(0.4, 0.6))

}
