% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwEvents.R
\name{pwEvents}
\alias{pwEvents}
\title{Expected number of events for a piecewise enrolment distribution}
\usage{
pwEvents(
  DCO = 40,
  f = fa,
  time = 1:5,
  cum.p = c(0.05, 0.15, 0.35, 0.65, 1),
  xi = rep(1, 5),
  drop.rate = 0.03/12,
  n = 300
)
}
\arguments{
\item{DCO}{Data cut-off time, calculated from first subject in.}

\item{f}{Density function of survival time}

\item{time}{A vector of time points to define the piecewise enrolment intervals
for example, time = c(1,2,3) defines enrolment intervals 0-1, 1-2, 2-3.}

\item{cum.p}{A vector of cumulative enrolment at each time point. It must have
the same length of time. For example, cum.p = c(0.2, 0.5, 1.0).}

\item{xi}{A vector of weights for each piecewise enrolment distribution function.
It must have the same length as time and cum.p. For example, xi = c(1,1,1)
means uniform enrollment in all intervals.}

\item{drop.rate}{Drop-Off rate in month. If the drop-off rate is 3\% per year,
then drop.rate = 0.03/12.}

\item{n}{Number of subjects}
}
\description{
This function calculates the expected number of events for piecewise enrolment
distribution.
}
\examples{

pwEvents(DCO = 40, f=fa, time=1:5, cum.p = c(0.05, 0.15, 0.35, 0.65, 1.0), 
xi=rep(1,5), drop.rate = 0.03/12, n = 300)

pwEvents(DCO = 40, f=fa, time=5, cum.p = 1, xi=2, drop.rate = 0.03/12, n = 300)


}
