% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePower.R
\name{simplePower}
\alias{simplePower}
\title{Power Calculation Based on Provided Number of Events And Rejection Boundary in Group Sequential Design}
\usage{
simplePower(
  events = c(126, 210),
  events0 = NULL,
  events1 = NULL,
  hr = 0.6,
  r = 0.5,
  bd.p = NULL,
  sf = gsDesign::sfLDOF,
  alpha = 0.025,
  variance = "H0"
)
}
\arguments{
\item{events}{A vector of target events for IAs and FA}

\item{events0}{A vector of target events for IAs and FA for control arm.
Required when variance option is H1.}

\item{events1}{A vector of target events for IAs and FA for experimental arm
Required when variance option is H1.}

\item{hr}{Hazard ratio (smaller better)}

\item{r}{Proportion of experimental arm subjects among all subjects}

\item{bd.p}{A vector of rejection bound in one-sided p value. When bd.p is
specified, sf and alpha are ignored.}

\item{sf}{Alpha spending function. Default sf = gsDesign::sfLDOF, i.e.,
LanDeMets implementation of O'Brien Fleming spending function.}

\item{alpha}{Overall alpha for the test in the group sequential design.
Default, alpha = 0.025. Alpha must be one-sided. When bd.p is specified, alpha
will be re-calculated. Required when bd.p is not specified.}

\item{variance}{Option for variance estimate. "H1" or "H0". Default H1, which is usually more conservative than H0.}
}
\value{
\itemize{
\item events     Target events at each analysis
\item bd         Rejection ounds at each analysis
\item marg.power Marginal power at each analysis
\item cum.power  Cumulative power by each analysis
\item overall.power Overall power
\item CV         Critical value (minimum detectable difference)
\item corr       Correlation matrix
}
}
\description{
This function calculates the power for the given number of events and rejection boundary
at each analysis for group sequential design
}
\examples{

#(1) O'Brien Fleming alpha spending, variance under H0
simplePower(events=c(126, 210), events0=NULL, events1=NULL, 
                        hr = 0.6, r = 0.5, 
                        bd.p=NULL, sf=gsDesign::sfLDOF, 
                        alpha=0.025, variance="H0")
#(2) O'Brien Fleming alpha spending, variance under H1
simplePower(events=c(126, 210), events0=c(66,115), events1=c(60,95), 
                        hr = 0.6, r = 0.5, 
                        bd.p=NULL, sf=gsDesign::sfLDOF, 
                        alpha=0.025, variance="H1")

#(3) Customized rejection bounds, variance under H1
simplePower(events=c(126, 210), events0=c(66,115), events1=c(60,95), 
                        hr = 0.6, r = 0.5, 
                        bd.p=c(0.004,0.024), sf=NULL, 
                        alpha=NULL, variance="H1")
simplePower(events=c(208, 287), events0=c(66,115), events1=c(60,95), 
                        hr = 0.6, r = 0.5, 
                        bd.p=c(0.016933/2,0.044898/2), sf=NULL, 
                        alpha=NULL, variance="H0")
simplePower(events=c(231, 287), events0=c(66,115), events1=c(60,95), 
                        hr = 0.6, r = 0.5, 
                        bd.p=c(0.016933/2,0.044898/2), sf=NULL, 
                        alpha=NULL, variance="H0")
                        
}
